requirements.txt
aes==1.2.0
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
cryptography==39.0.1
Django==4.1.6
django-cors-headers==3.13.0
docopt==0.6.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
Naked==0.1.32
openai==0.27.2
pipreqs==0.4.11
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
PyMySQL==1.0.2
PyYAML==6.0
requests==2.28.2
shellescape==3.8.1
sqlparse==0.4.3
tqdm==4.64.1
tzdata==2022.7
urllib3==1.26.14
wenxin-api==0.2.1
xlwt==1.3.0
yarg==0.1.9
yarl==1.8.2